+++ src/components/QueuePanel.tsx
@@ -5,13 +5,12 @@ import { registerQueueListScrollTopReader, consumePendingQueueListScrollTop } fr
 import { songToTrack } from '../utils/songToTrack';
 import type { Track } from '../store/playerStoreTypes';
 import React, { useState, useRef, useMemo, useEffect, useLayoutEffect } from 'react';
-import { createPortal } from 'react-dom';
 import { usePlayerStore } from '../store/playerStore';
 import { useOrbitStore } from '../store/orbitStore';
 import OrbitGuestQueue from './OrbitGuestQueue';
 import OrbitQueueHead from './OrbitQueueHead';
 import HostApprovalQueue from './HostApprovalQueue';
-import { Play, Music, Trash2, Save, FolderOpen, Shuffle, Infinity, Waves, MicVocal, ListMusic, Check, MoveRight, Radio, HardDrive, ChevronDown, Info, Share2 } from 'lucide-react';
+import { Play, Trash2, Save, FolderOpen, Shuffle, Infinity, Waves, MicVocal, ListMusic, Check, MoveRight, Radio, Info, Share2 } from 'lucide-react';
 import { usePlaylistStore } from '../store/playlistStore';
 import { useCachedUrl } from './CachedImage';
 import { useTranslation } from 'react-i18next';
@@ -29,17 +28,14 @@ import { TFunction } from 'i18next';
 import OverlayScrollArea from './OverlayScrollArea';
 import { useLuckyMixStore } from '../store/luckyMixStore';
 import { useQueueToolbarStore, QueueToolbarButtonId } from '../store/queueToolbarStore';
-import { loudnessGainPlaceholderUntilCacheDb } from '../utils/loudnessPlaceholder';
-import { effectiveLoudnessPreAnalysisAttenuationDb } from '../utils/loudnessPreAnalysisSlider';
 import {
   DurationMode,
   formatTime,
-  formatQueueReplayGainParts,
-  renderStars,
 } from '../utils/queuePanelHelpers';
 import { SavePlaylistModal } from './queuePanel/SavePlaylistModal';
 import { LoadPlaylistModal } from './queuePanel/LoadPlaylistModal';
 import { QueueHeader } from './queuePanel/QueueHeader';
+import { QueueCurrentTrack } from './queuePanel/QueueCurrentTrack';
 
 export default function QueuePanel() {
   const orbitRole = useOrbitStore(s => s.role);
@@ -457,218 +453,30 @@ function QueuePanelHostOrSolo() {
       />
 
       {currentTrack && !isNowPlayingCollapsed && (
-        <div className="queue-current-track">
-          {(() => {
-            const baseParts = [
-              currentTrack.suffix?.toUpperCase(),
-              currentTrack.bitRate ? `${currentTrack.bitRate} kbps` : undefined,
-              (() => {
-                const bd = currentTrack.bitDepth;
-                const sr = currentTrack.samplingRate ? `${currentTrack.samplingRate / 1000} kHz` : '';
-                if (bd && sr) return `${bd}/${sr}`;
-                if (bd) return `${bd}-bit`;
-                if (sr) return sr;
-                return undefined;
-              })(),
-            ].filter(Boolean) as string[];
-            const rgParts = formatQueueReplayGainParts(currentTrack, t);
-            const baseLine = baseParts.join(' · ');
-            const rgLine = rgParts.join(' · ');
-            const isLoudnessActive = normalizationEngine === 'loudness' || normalizationEngineLive === 'loudness';
-            const liveGainLabel = (() => {
-              if (normalizationNowDb != null && Number.isFinite(normalizationNowDb)) {
-                return `${normalizationNowDb >= 0 ? '+' : ''}${normalizationNowDb.toFixed(2)} dB`;
-              }
-              if (isLoudnessActive && Number.isFinite(loudnessPreAnalysisAttenuationDb)) {
-                const preEff = effectiveLoudnessPreAnalysisAttenuationDb(
-                  loudnessPreAnalysisAttenuationDb,
-                  authLoudnessTargetLufs,
-                );
-                const ph = loudnessGainPlaceholderUntilCacheDb(
-                  authLoudnessTargetLufs,
-                  preEff,
-                );
-                return `${ph >= 0 ? '+' : ''}${ph.toFixed(2)} dB`;
-              }
-              return '—';
-            })();
-            const tgtNum = normalizationTargetLufs ?? authLoudnessTargetLufs;
-            const targetLabel = `${tgtNum} LUFS`;
-            if (!baseLine && !rgLine && !playbackSource) return null;
-            const showRgLine = !isLoudnessActive && expandReplayGain && !!rgLine;
-            const showLufsLine = isLoudnessActive && expandReplayGain;
-            return (
-              <div className={`queue-current-tech${showRgLine ? ' queue-current-tech--two-line' : ''}`}>
-                <div className="queue-current-tech-stack">
-                  <div className="queue-current-tech-row">
-                    {playbackSource && (
-                      <span
-                        className="queue-current-tech-source"
-                        data-tooltip={
-                          playbackSource === 'offline'
-                            ? t('queue.sourceOffline')
-                            : playbackSource === 'hot'
-                              ? t('queue.sourceHot')
-                              : t('queue.sourceStream')
-                        }
-                        aria-hidden
-                      >
-                        {playbackSource === 'offline' && <FolderOpen size={11} strokeWidth={2.25} />}
-                        {playbackSource === 'hot' && <HardDrive size={11} strokeWidth={2.25} />}
-                        {playbackSource === 'stream' && <Waves size={11} strokeWidth={2.25} />}
-                      </span>
-                    )}
-                    {baseLine && <span className="queue-current-tech-main">{baseLine}</span>}
-                    {!isLoudnessActive && rgLine && (
-                      <button
-                        type="button"
-                        className={`queue-current-tech-rg-badge${showRgLine ? ' queue-current-tech-rg-badge--open' : ''}`}
-                        data-tooltip={`${t('queue.replayGain')} · ${rgLine}`}
-                        aria-expanded={showRgLine}
-                        aria-label={t('queue.replayGain')}
-                        onClick={() => setExpandReplayGain(!expandReplayGain)}
-                      >
-                        RG
-                        <ChevronDown size={9} strokeWidth={2.5} />
-                      </button>
-                    )}
-                    {isLoudnessActive && (
-                      <button
-                        type="button"
-                        className={`queue-current-tech-rg-badge${showLufsLine ? ' queue-current-tech-rg-badge--open' : ''}`}
-                        data-tooltip={`LUFS · ${liveGainLabel} · TGT · ${targetLabel}`}
-                        aria-expanded={showLufsLine}
-                        aria-label="LUFS"
-                        onClick={() => setExpandReplayGain(!expandReplayGain)}
-                      >
-                        LUFS
-                        <ChevronDown size={9} strokeWidth={2.5} />
-                      </button>
-                    )}
-                  </div>
-                  {showRgLine && (
-                    <span className="queue-current-tech-rg">
-                      <span className="queue-current-tech-rg-label">{t('queue.replayGain')}</span>
-                      {' · '}{rgLine}
-                    </span>
-                  )}
-                  {showLufsLine && (
-                    <span className="queue-current-tech-rg">
-                      <span className="queue-current-tech-rg-label">Loudness</span>
-                      {' · '}
-                      <button
-                        type="button"
-                        className="queue-current-tech-metric queue-current-tech-metric--lufs-reanalyze"
-                        onClick={e => {
-                          e.stopPropagation();
-                          setLufsTgtOpen(false);
-                          void reanalyzeLoudnessForTrack(currentTrack.id);
-                        }}
-                        data-tooltip={t('queue.clearCachedLoudnessWaveform')}
-                        aria-label={t('queue.clearCachedLoudnessWaveform')}
-                      >
-                        {liveGainLabel}
-                      </button>
-                      {' · '}
-                      <span className="queue-current-tech-rg-label">TGT</span>
-                      {' · '}
-                      <button
-                        type="button"
-                        ref={lufsTgtBtnRef}
-                        className="queue-current-tech-metric"
-                        onClick={e => {
-                          e.stopPropagation();
-                          setLufsTgtOpen(v => !v);
-                        }}
-                        data-tooltip="Change target integrated loudness"
-                        aria-haspopup="listbox"
-                        aria-expanded={lufsTgtOpen}
-                      >
-                        {targetLabel}
-                      </button>
-                      {lufsTgtOpen &&
-                        createPortal(
-                          <div
-                            ref={lufsTgtMenuRef}
-                            className="queue-lufs-tgt-menu"
-                            style={{
-                              ...lufsTgtPopStyle,
-                              background: 'var(--bg-card)',
-                              border: '1px solid var(--border, rgba(255,255,255,0.12))',
-                              borderRadius: 8,
-                              boxShadow: '0 6px 24px rgba(0,0,0,0.35)',
-                              padding: 6,
-                              overflow: 'auto',
-                            }}
-                            role="listbox"
-                            aria-label="LUFS target"
-                            onClick={e => e.stopPropagation()}
-                          >
-                            {([-10, -12, -14, -16] as const).map((v) => (
-                              <button
-                                key={v}
-                                type="button"
-                                onClick={e => {
-                                  e.stopPropagation();
-                                  if (v !== authLoudnessTargetLufs) {
-                                    setLoudnessTargetLufs(v);
-                                  }
-                                  setLufsTgtOpen(false);
-                                }}
-                                style={{
-                                  display: 'block',
-                                  width: '100%',
-                                  textAlign: 'left',
-                                  padding: '6px 8px',
-                                  borderRadius: 6,
-                                  border: 'none',
-                                  background: v === authLoudnessTargetLufs ? 'color-mix(in srgb, var(--accent) 18%, transparent)' : 'transparent',
-                                  color: 'var(--text-primary)',
-                                  cursor: 'pointer',
-                                  font: 'inherit',
-                                }}
-                              >
-                                {v} LUFS
-                              </button>
-                            ))}
-                          </div>,
-                          document.body,
-                        )}
-                    </span>
-                  )}
-                </div>
-              </div>
-            );
-          })()}
-          <div className="queue-current-track-body">
-            <div className="queue-current-cover">
-              {currentTrack.coverArt ? (
-                <img src={currentCoverSrc} alt="" loading="eager" />
-              ) : (
-                <div className="fallback"><Music size={32} /></div>
-              )}
-            </div>
-            <div className="queue-current-info">
-              <h3 className="truncate">{currentTrack.title}</h3>
-              <div
-                className={`queue-current-sub truncate${currentTrack.artistId ? ' is-link' : ''}`}
-                onClick={() => currentTrack.artistId && navigate(`/artist/${currentTrack.artistId}`)}
-              >{currentTrack.artist}</div>
-              <div
-                className={`queue-current-sub truncate${currentTrack.albumId ? ' is-link' : ''}`}
-                onClick={() => currentTrack.albumId && navigate(`/album/${currentTrack.albumId}`)}
-              >{currentTrack.album}</div>
-              {currentTrack.year && (
-                <div className="queue-current-sub">{currentTrack.year}</div>
-              )}
-              {(() => {
-                const label = orbitAttributionLabel(currentTrack.id);
-                return label ? <div className="queue-current-sub queue-current-attribution">{label}</div> : null;
-              })()}
-              {renderStars(userRatingOverrides[currentTrack.id] ?? currentTrack.userRating)}
-            </div>
-          </div>
-        </div>
+        <QueueCurrentTrack
+          currentTrack={currentTrack}
+          currentCoverSrc={currentCoverSrc}
+          userRatingOverrides={userRatingOverrides}
+          orbitAttributionLabel={orbitAttributionLabel}
+          navigate={navigate}
+          playbackSource={playbackSource}
+          normalizationEngine={normalizationEngine}
+          normalizationEngineLive={normalizationEngineLive}
+          normalizationNowDb={normalizationNowDb}
+          normalizationTargetLufs={normalizationTargetLufs}
+          authLoudnessTargetLufs={authLoudnessTargetLufs}
+          loudnessPreAnalysisAttenuationDb={loudnessPreAnalysisAttenuationDb}
+          expandReplayGain={expandReplayGain}
+          setExpandReplayGain={setExpandReplayGain}
+          reanalyzeLoudnessForTrack={reanalyzeLoudnessForTrack}
+          setLoudnessTargetLufs={setLoudnessTargetLufs}
+          lufsTgtOpen={lufsTgtOpen}
+          setLufsTgtOpen={setLufsTgtOpen}
+          lufsTgtBtnRef={lufsTgtBtnRef}
+          lufsTgtMenuRef={lufsTgtMenuRef}
+          lufsTgtPopStyle={lufsTgtPopStyle}
+          t={t}
+        />
       )}
 
       {activeTab === 'queue' ? (<>
